Added "preIntersectionStyle" prop to ViewportImage
Implemented an optional "preIntersectionStyle" prop to the ViewportImage component which is added to the class list immediately before both the "placeholder" is replaced by the "src" and the "intersectionStyle" is added to the class list.  A "preIntersectionStyle" argument must be accompanied by an "intersectionStyle" argument or it will be ignored.  However, an "intersectionStyle" argument may function on its own.  Both "preIntersectionStyle" and "intersectionStyle" are removed from the class list after the "intersectionStyle" animation is complete.