Added "active" state to Item documents
Removed the ability to edit the "name" of Item documents from the UI in order to maintain the integrity of past voting results.  However, to nondestructively simulate deleting Item documents or editing the "name" property of an Item document from the UI, an "active" state has been added to the Item document model that is controlled by a toggle switch in the EditItem component.  Inactive Item documents will not appear in the list of candidates and Item documents can neither be added nor edited while a vote is in progress.