Updated Portal and ResultDetail components
Added "okCallback" and "closeCallback" function props to the Portal component that are called when a "keydown" event targeting the Enter and Escape keys is dispatched.  Additionally, added new label mixins to unionize the entry label styles in the ResultDetailTableItemRow and ItemDetail components and fixed the ResultDetailTableUser component styling so that it abides by the "RESULTS_TABLE_USER_HEIGHT" constant.