Updated styling for Vote and List components
Implemented new styling for both the Vote and List components, including a responsive design for a compact layout to support smaller screens.  Additionally, Candidate items dropped onto the Votes list will be returned if the Votes list is already populated with the maximum amount of rank selections for the active vote.